main.py
import os
import logging
from datetime import datetime

# Setup logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def load_data():
    """
    Function to load data from the IRS Form 990 dataset.
    This is a placeholder and should be implemented with actual data loading logic.
    """
    logger.info("Loading data...")
    # Implement your data loading logic here
    return "Data loaded successfully"

def process_data(data):
    """
    Function to process the loaded data.
    This is a placeholder and should be implemented with your data processing logic.
    """
    logger.info("Processing data...")
    # Implement your data processing logic here
    return "Data processed successfully"

def predict_financial_health(processed_data):
    """
    Function to predict financial health based on processed data.
    This is a placeholder and should be implemented with your prediction model.
    """
    logger.info("Predicting financial health...")
    # Implement your prediction logic here
    return "Financial health prediction complete"

def main():
    """
    Main function to orchestrate the nonprofit financial health prediction process.
    """
    logger.info(f"Starting Nonprofit Financial Health Predictor at {datetime.now()}")
    
    try:
        data = load_data()
        processed_data = process_data(data)
        prediction = predict_financial_health(processed_data)
        
        logger.info(f"Prediction process completed successfully: {prediction}")
    except Exception as e:
        logger.error(f"An error occurred: {str(e)}")

if __name__ == "__main__":
    main()